SME: Widen the whilelt arguments to a fixed-width type
c1656e701 turned sme_traits::whilelt into a template deduced from its
arguments, dropping the casts that used to reach svwhilelt_b32/_b64. The ACLE
overloads those on the fixed-width integer types only -- int32_t, int64_t,
uint32_t, uint64_t -- so an argument whose type merely has the same width as
one of them matches none exactly and the call is ambiguous. The packers pass
Index, i.e. long: that is int64_t on Linux, and a distinct type wherever
int64_t is long long, arm64 macOS among them. There every SME build fails, a
plain MatrixXf product included.

Cast in the two traits rather than at each caller, to int64_t: signed, like the
Index the callers pass, and wide enough for all of them.

Reproduced and fixed on Apple M4 hardware (SVL=512, FEAT_SME_F64F64,
-mcpu=apple-m4), where the parent commit does not compile. product_sme 1-3 pass,
as do the 137 subtests of product_small, product_large, product_extra,
product_mmtr, product_symm, product_syrk, product_selfadjoint, product_trmm,
product_trmv, product_trsolve and product_notemporary.

+++ Eigen/src/Core/arch/SME/GeneralBlockPanelKernel.h
@@ -84,9 +84,13 @@ struct sme_traits<float> {
   using Vec2 = svfloat32x2_t;
   using Vec4 = svfloat32x4_t;
   static EIGEN_ALWAYS_INLINE int svl() __arm_streaming_compatible { return static_cast<int>(svcntsw()); }
+  // svwhilelt is overloaded on the fixed-width integer types only, so widen
+  // here: an argument of a distinct type of the same width -- `Index` where
+  // `int64_t` is `long long` -- would otherwise match none of them exactly and
+  // make the call ambiguous.
   template <typename T, typename = std::enable_if_t<std::is_integral<T>::value>>
   static EIGEN_ALWAYS_INLINE svbool_t whilelt(T begin, T end) __arm_streaming {
-    return svwhilelt_b32(begin, end);
+    return svwhilelt_b32(static_cast<int64_t>(begin), static_cast<int64_t>(end));
   }
   static EIGEN_ALWAYS_INLINE svbool_t ptrue() __arm_streaming { return svptrue_b32(); }
   static EIGEN_ALWAYS_INLINE svcount_t ptrue_c() __arm_streaming { return svptrue_c32(); }
@@ -102,7 +106,7 @@ struct sme_traits<double> {
   static EIGEN_ALWAYS_INLINE int svl() __arm_streaming_compatible { return static_cast<int>(svcntsd()); }
   template <typename T, typename = std::enable_if_t<std::is_integral<T>::value>>
   static EIGEN_ALWAYS_INLINE svbool_t whilelt(T begin, T end) __arm_streaming {
-    return svwhilelt_b64(begin, end);
+    return svwhilelt_b64(static_cast<int64_t>(begin), static_cast<int64_t>(end));
   }
   static EIGEN_ALWAYS_INLINE svbool_t ptrue() __arm_streaming { return svptrue_b64(); }
   static EIGEN_ALWAYS_INLINE svcount_t ptrue_c() __arm_streaming { return svptrue_c64(); }
